Used a single dialog component for delete Warnings
There are now only two files, a generic delete warning and another warning for deleting a whole scenerio, this is more efficient than the last implementation.
