+++ Ghost_in_the_Shell/client_folder_react_app/ethisim/src/components/EditorComponents/ConversationEditorComponents/StakeHoldersComponent/StakeHolderQuestions/question.js
@@ -5,7 +5,7 @@ import { Button } from '@material-ui/core';
 import Box from '@material-ui/core/Box';
 import htmlToText from 'html-to-text';
 import PropTypes from 'prop-types';
-import DeletePopupConversationQuestions from '../../../../DeleteWarnings/DeletePopupConversationQuestions';
+import GenericDeleteWarning from '../../../../DeleteWarnings/GenericDeleteWarning';
 
 const useStyles = makeStyles((theme) => ({
     margin: {
@@ -65,6 +65,12 @@ export default function QuestionField({
         );
     };
 
+    //Used for delete Warning Popup window
+    const [open, setOpen] = React.useState(false);
+    const handleClickOpen = () => {
+        setOpen(true);
+    };
+
     // eslint-disable-next-line
     let handleChange = (content, editor) => {
         //TODO Implement
@@ -121,8 +127,19 @@ export default function QuestionField({
                             Save
                         </Button>
                     </div>
-                    <DeletePopupConversationQuestions
+                    <Button
+                        className={classes.margin}
+                        variant="contained"
+                        color="primary"
+                        onClick={handleClickOpen}
+                    >
+                        Delete
+                    </Button>
+
+                    <GenericDeleteWarning
                         remove={() => removeQuestion(id)}
+                        open={open}
+                        setOpen={setOpen}
                     />
                 </Box>
             </Box>
