+++ Ghost_in_the_Shell/client_folder_react_app/ethisim/src/components/EditorComponents/LogisticsPageComponents/Logistics.js
@@ -4,7 +4,6 @@ import AuthorField from './Author';
 import { isBrowser } from 'react-device-detect';
 import { Button, TextField, Typography, Container } from '@material-ui/core';
 import { mockUnfinishedScenario } from '../../../shared/mockScenarioData';
-import DeletePopupScenarioLogistics from '../../DeleteWarnings/DeletePopupScenarioLogistics';
 
 const useStyles = makeStyles((theme) => ({
     textfields: {
@@ -161,7 +160,6 @@ export default function Logistics(props) {
                         View Version History
                     </Button>
                 </form>
-                <DeletePopupScenarioLogistics />
             </div>
         </Container>
     );
@@ -239,7 +237,6 @@ export default function Logistics(props) {
                             View Version History
                         </Button>
                     </form>
-                    <DeletePopupScenarioLogistics />
                 </div>
                 <Button
                     className={classes.saveButton}
